fix(TopologyResourcePanel): fix bad import
