+++ frontend/packages/knative-plugin/src/components/overview/KnativeResourceOverviewPage.tsx
@@ -1,17 +1,23 @@
 import * as React from 'react';
+import { connect } from 'react-redux';
 import { Kebab, LoadingBox } from '@console/internal/components/utils';
 import { ResourceOverviewDetails } from '@console/internal/components/overview/resource-overview-details';
-import { OverviewItem } from '@console/shared';
 import { groupVersionFor, K8sKind } from '@console/internal/module/k8s';
-import { getKsResourceModel } from '../../utils/get-knative-resources';
-import { KnativeOverview } from './KnativeOverview';
+import { RootState } from '@console/internal/redux';
+import { OverviewItem } from '@console/shared';
+import { KNATIVE_SERVING_APIGROUP, KNATIVE_EVENT_SOURCE_APIGROUP } from '../../const';
 import OverviewDetailsKnativeResourcesTab from './OverviewDetailsKnativeResourcesTab';
+import KnativeOverview from './KnativeOverview';
 
-export type KnativeResourceOverviewPageProps = {
-  item?: OverviewItem;
-  knativeModels?: K8sKind[];
+interface StateProps {
   kindsInFlight?: boolean;
-};
+  knativeModels?: K8sKind[];
+}
+
+export interface KnativeResourceOverviewPageProps extends StateProps {
+  item?: OverviewItem;
+}
+
 const tabs = [
   {
     name: 'Overview',
@@ -48,4 +54,17 @@ export const KnativeResourceOverviewPage: React.ComponentType<KnativeResourceOve
   );
 };
 
-export default getKsResourceModel(KnativeResourceOverviewPage);
+const mapStateToProps = (state: RootState): StateProps => {
+  return {
+    kindsInFlight: state.k8s.getIn(['RESOURCES', 'inFlight']),
+    knativeModels: state.k8s
+      .getIn(['RESOURCES', 'models'])
+      .filter(
+        (model: K8sKind) =>
+          model.apiGroup === KNATIVE_SERVING_APIGROUP ||
+          model.apiGroup === KNATIVE_EVENT_SOURCE_APIGROUP,
+      ),
+  };
+};
+
+export default connect(mapStateToProps)(KnativeResourceOverviewPage);
