+++ frontend/packages/knative-plugin/src/components/overview/__tests__/KnativeOverview.spec.tsx
@@ -5,7 +5,7 @@ import { PodRing, OverviewItem } from '@console/shared';
 import { ResourceSummary } from '@console/internal/components/utils';
 import { revisionObj } from '@console/dev-console/src/components/topology/__tests__/topology-knative-test-data';
 import { RevisionModel } from '../../../models';
-import { KnativeOverview } from '../KnativeOverview';
+import KnativeOverview from '../KnativeOverview';
 
 describe('KnativeOverview', () => {
   let item: OverviewItem;
